✨ feat(codec): modified order column type

+++ database/migrations/2022_06_08_141634_create_codec_videos_table.php
@@ -15,7 +15,7 @@ return new class extends Migration {
       $table->id();
 
       $table->string('codec', 16);
-      $table->integer('order')->nullable();
+      $table->tinyInteger('order')->nullable();
 
       $table->timestamps();
     });
